move calculate week status util to utls
