+++ packages/psss/src/api/queries/useCountryWeeklyReport.js
@@ -5,32 +5,10 @@
 
 import { useState } from 'react';
 import keyBy from 'lodash.keyby';
-import isequal from 'lodash.isequal';
 import { useTableData } from './useTableData';
-import { subtractWeeksFromPeriod } from '../../utils';
+import { subtractWeeksFromPeriod, calculateWeekStatus } from '../../utils';
 import { REPORT_STATUSES } from '../../constants';
 import { useUpcomingReport } from './useUpcomingReport';
-
-const extractReportData = report => {
-  const { AFR, DIA, DLI, ILI, PF } = report;
-  return { AFR, DIA, DLI, ILI, PF };
-};
-
-export const reportsAreEqual = (reportA, reportB) => {
-  const a = extractReportData(reportA);
-  const b = extractReportData(reportB);
-  return isequal(a, b);
-};
-
-export const calculateWeekStatus = (report, confirmedReport) => {
-  if (!confirmedReport || !report) {
-    return REPORT_STATUSES.OVERDUE;
-  }
-
-  return reportsAreEqual(report, confirmedReport)
-    ? REPORT_STATUSES.SUBMITTED
-    : REPORT_STATUSES.RESUBMIT;
-};
 
 /**
  * Get the weekly reports data by combining the un-confirmed and confirmed weekly reports
