+++ packages/psss/src/utils/__tests/calculateWeekStatus.test.js
@@ -3,12 +3,12 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 import keyBy from 'lodash.keyby';
-import { REPORT_STATUSES } from '../../../constants';
-import { reportsAreEqual, calculateWeekStatus } from '../useCountryWeeklyReport';
+import { REPORT_STATUSES } from '../../constants';
+import { reportsAreEqual, calculateWeekStatus } from '../calculateWeekStatus';
 import unconfirmedData from './fixtures/unConfirmedReport.json';
 import confirmedData from './fixtures/confirmedReport.json';
 
-describe('useCountryWeeklyReport', () => {
+describe('calculateWeekStatus', () => {
   const reportsByPeriod = keyBy(unconfirmedData.data.results, 'period');
   const confirmedReportsByPeriod = keyBy(confirmedData.data.results, 'period');
