+++ packages/psss/src/utils/index.js
@@ -5,3 +5,4 @@
 
 export * from './countries';
 export * from './period';
+export * from './calculateWeekStatus';
